Create and Complete Tomato Soup Recipe Page
Create new html page for tomato soup recipe.
Filled in the page with description, ingredients,
and instructions for cooking. Imported a tomato
soup image into the img folder for use on the soup
recipe page. Connect the index page to the tomato
soup page. Include two new sections for recipes on
the index page, soup/stews and desserts.